usb keyboard client, serial settings console and oled status display
Adds a USB keyboard to Dreamcast keyboard bridge on top of DreamPicoPort,
plus a serial configuration path and an optional status screen.

Keyboard client. Upstream shipped a complete DreamcastKeyboard maple
implementation that nothing ever instantiated, and a usb host layer that
handled gamepads only. This is the glue between them, following the
existing GamepadHost pattern: hid_app fills a report struct on core 1,
DreamcastKeyboard serves it from core 0 on GET_CONDITION. No usage code
translation is needed, since the Dreamcast reports the same HID codes a usb
keyboard does and the modifier byte matches on bits 0 through 6. Lock LED
state is pushed back to the keyboard, and the declared layout is
configurable for japanese releases that expect 106 keys.

Settings over serial. Configuration was reachable only over WebUSB, so a
browser was the only way to set up a device. The same fields now live on
the tty parser under =, as line oriented text that is typeable by hand and
parseable by a script. The WebUSB handler is untouched.

OLED status display. Optional SSD1306 on I2C1 with three pages cycled by a
button: typing stats, a debug page, and a title card. The panel is owned
entirely by core 1 and only changed pages are written, because a full frame
is about 25 ms of blocking i2c and the console's poll window is ~16 ms.

bench-host. The host firmware with the transceiver direction pin disabled,
so two bare RP2040s can be wired together and a client tested without a
console.

Build fixes. The project did not compile against pico-sdk 2.3.0 with a
current ARM toolchain: gcc 15 tightened transitive includes, tinyusb 0.18
removed tud_vendor_n_write_clear, and -Werror=narrowing rejects a
strlen()+1 into a uint16_t. Also fixes a signed overflow in
DreamcastKeyboard's condition response, where uint8_t operands promote to
int and shifting 0x80 left by 24 is undefined. Reachable in normal use, and
never caught because nothing had instantiated the class.

The WebUSB landing page points at this fork and its browser announcement
defaults off, since configuration happens over serial here.

Verified on real hardware: a console assigns a player slot and polls the
adapter, keystrokes cross from usb to maple, modifiers and five key
rollover work. Declared current draw is still a guess and the device info
version string is invented.

+++ src/hostLib/webusb_parsers/FlycastWebUsbCommandHandler.cpp
@@ -22,6 +22,7 @@
 // SOFTWARE.
 
 #include "FlycastWebUsbCommandHandler.hpp"
+#include <cstdint>
 
 #include "hal/MapleBus/MaplePacket.hpp"
 #include "hal/System/LockGuard.hpp"
@@ -142,7 +143,7 @@ void FlycastWebUsbCommandHandler::process(
             char buffer[mIdentification.getSerialSize() + 1] = {0};
             mIdentification.getSerial(buffer, sizeof(buffer) - 1);
             buffer[sizeof(buffer) - 1] = '\0';
-            responseFn(kResponseSuccess, {{buffer, strlen(buffer) + 1}});
+            responseFn(kResponseSuccess, {{buffer, static_cast<uint16_t>(strlen(buffer) + 1)}});
         }
         return;
