+++ src/hostLib/webusb_parsers/SystemWebUsbCommandHandler.cpp
@@ -22,6 +22,7 @@
 // SOFTWARE.
 
 #include "SystemWebUsbCommandHandler.hpp"
+#include <cstdint>
 
 #include <cstring>
 
@@ -69,7 +70,7 @@ void SystemWebUsbCommandHandler::process(
             char buffer[mIdentification.getSerialSize() + 1] = {0};
             mIdentification.getSerial(buffer, sizeof(buffer) - 1);
             buffer[sizeof(buffer) - 1] = '\0';
-            responseFn(kResponseSuccess, {{buffer, strlen(buffer) + 1}});
+            responseFn(kResponseSuccess, {{buffer, static_cast<uint16_t>(strlen(buffer) + 1)}});
         }
         return;
